Move API types into app/types and update imports
